requirements.txt
Flask
Flask-Mobility
Flask-SocketIO
gevent
gevent-websocket
python-dotenv
requests
pocketbase==0.10.0